Create SOC-SOH monitoring though arduino
The above code uses coulomb counting method for the estimation of SOC of a lithium ion cell(s). The values for end of charge and end of discharge voltage are to be taken as per cell chemistry and number of cells in series. I'll bring in further modifications to this code and make it more robust and accurate.